Make it better
